Moved the metric data class into the metric module
