+++ metric_system/functions/metric.py
@@ -1,5 +1,7 @@
 from abc import ABC
+from dataclasses import dataclass
 from datetime import datetime
+from typing import Any
 from mypy_boto3_cloudwatch.literals import StandardUnitType
 
 
@@ -39,3 +41,59 @@ class Metric(ABC):
     @property
     def metric_init_time(self) -> datetime:
         return self._metric_init
+
+
+@dataclass
+class DataPoint:
+    timestamp: datetime
+    unit: StandardUnitType
+    value: float
+
+    @classmethod
+    def create_from(cls, metric: Metric):
+        return DataPoint(datetime.now(), metric.unit, metric.value)
+
+
+@dataclass
+class MetricJsonData:
+    metric_name: str
+    data_points: list[DataPoint]
+
+    @classmethod
+    def from_json(cls, obj: Any):
+        "Parses the JSON representation of this class into an instance of this class"
+        # Ensure that we have our attributes
+        if not "metric_name" in obj or not "data_points" in obj:
+            raise TypeError("Error deserializing metric data!")
+        metric_name = obj["metric_name"]
+        data_points_json = obj["data_points"]
+
+        # Maps data point json => data point object, erroring out if needed
+        def map_data_points(data_point: Any) -> DataPoint:
+            if not all(
+                ["timestamp" in data_point, "unit" in data_point, "value" in data_point]
+            ):
+                raise TypeError("Error deserializing metric data!")
+            timestamp = datetime.strptime(
+                data_point["timestamp"], "%Y-%m-%d %H:%M:%S.%f"
+            )
+            unit = data_point["unit"]
+            value = float(data_point["value"])
+            return DataPoint(timestamp, unit, value)
+
+        # Call the mapping function
+        data_points = list(map(map_data_points, data_points_json))
+        return MetricJsonData(metric_name, data_points)
+
+    def to_json(self) -> Any:
+        "Makes this object JSON-serializable by treating the data points as dicts"
+
+        def data_point_to_json(data_point: DataPoint):
+            return {
+                "timestamp": str(data_point.timestamp),
+                "unit": data_point.unit,
+                "value": data_point.value,
+            }
+
+        data_points = list(map(data_point_to_json, self.data_points))
+        return {"metric_name": self.metric_name, "data_points": data_points}
